# api/openapi.yaml
openapi: 3.0.4
info:
  title: Dployr API
  description: |
    Your app, your server, your rules!
    
    ## Authentication
    Dployr instances use JWT-based authentication with tokens **issued by base**.

    ### Token Types (issued by base)
    - **Access Token**: Short-lived (e.g. 10 minutes) for API requests to instances
    - **Refresh Token**: Long-lived (24+ hours) for obtaining new access tokens from base

    ### Authentication Flow
    See https://docs.dployr.dev/auth for more details.
    
    ## Authorization
    Role-based access control with four levels:
    - **Owner**: Full system control, can manage admins
    - **Admin**: Infrastructure management, user management, secrets
    - **Developer**: Deploy apps, view logs and events
    - **Viewer**: Read-only access to services

    See https://docs.dployr.dev/auth for more details.
  version: 1.1.1
  contact:
    name: Emmanuel Madehin
    email: hello@dployr.dev
  license:
    name: MIT
    url: https://raw.githubusercontent.com/dployr-io/dployr/refs/heads/master/LICENSE
servers:
  - url: "{address}"
    description: Configurable server
    variables:
      address:
        default: http://localhost:7879
        description: Instance URL
security:
  - BearerAuth: []
paths:
  /deployments:
    get:
      tags:
        - Deployments
      summary: List deployments
      description: Retrieve a list of deployments (Developer+ required)
      operationId: listDeployments
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of deployments to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of deployments to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of deployments
          content:
            application/json:
              schema:
                type: object
                properties:
                  deployments:
                    type: array
                    items:
                      $ref: '#/components/schemas/Deployment'
                  total:
                    type: integer
                    description: Total number of deployments
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Deployments
      summary: Create deployment
      description: Create a new deployment (Developer+ required)
      operationId: createDeployment
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeployRequest'
      responses:
        '201':
          description: Deployment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeployResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /services:
    get:
      tags:
        - Services
      summary: List services
      description: Retrieve a list of services (Viewer+ required)
      operationId: listServices
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of services to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of services to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of services
          content:
            application/json:
              schema:
                type: object
                properties:
                  services:
                    type: array
                    items:
                      $ref: '#/components/schemas/Service'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /services/{serviceId}:
    get:
      tags:
        - Services
      summary: Get service
      description: Retrieve a specific service by ID (Viewer+ required)
      operationId: getService
      security:
        - BearerAuth: []
      parameters:
        - name: serviceId
          in: path
          required: true
          description: Service ID
          schema:
            type: string
      responses:
        '200':
          description: Service details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Services
      summary: Update service
      description: Update a service configuration (Developer+ required)
      operationId: updateService
      security:
        - BearerAuth: []
      parameters:
        - name: serviceId
          in: path
          required: true
          description: Service ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceUpdate'
      responses:
        '200':
          description: Service updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags:
        - Services
      summary: Partially update service
      description: Partially update a service configuration (Developer+ required)
      operationId: patchService
      security:
        - BearerAuth: []
      parameters:
        - name: serviceId
          in: path
          required: true
          description: Service ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceUpdate'
      responses:
        '200':
          description: Service updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Services
      summary: Delete service
      description: Delete a service (Admin+ required)
      operationId: deleteService
      security:
        - BearerAuth: []
      parameters:
        - name: serviceId
          in: path
          required: true
          description: Service ID
          schema:
            type: string
      responses:
        '204':
          description: Service deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /logs/stream:
    get:
      tags:
        - Logs
      summary: Open log stream
      description: |
        Open a keep alive connection to stream logs via server sent events (SSE) (Developer+ required).
        
        - `data:` events contain log lines
        - `event: error` indicates streaming errors
        - `event: done` indicates stream completion
        - `: heartbeat` every 15 seconds to keep the connection alive
      operationId: openLogStream
      security:
        - BearerAuth: []
      parameters:
        - name: token
          in: query
          required: true
          description: Bearer token for authentication
          schema:
            type: string
        - name: id
          in: query
          required: true
          description: Service or deployment ID to stream logs for
          schema:
            type: string
      responses:
        '200':
          description: Logs stream is successfully established
          content:
            text/event-stream:
              schema:
                type: string
                description: Server-Sent Events stream with log data
                example: |
                  data: cloning repository
                  
                  data: installing nodejs@18.1.0
                  
                  : heartbeat
                  
                  event: done
                  data: stream completed
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '405':
          description: Method not allowed (only GET is supported)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Runtime:
      type: string
      enum:
        - static
        - golang
        - php
        - python
        - nodejs
        - ruby
        - dotnet
        - java
        - docker
        - k3s
        - custom
    RuntimeObj:
      type: object
      required:
        - type
      properties:
        type:
          $ref: '#/components/schemas/Runtime'
        version:
          type: string
          example: "18.0.0"
    RemoteObj:
      type: object
      properties:
        url:
          type: string
          format: uri
          example: "https://github.com/user/repo.git"
        branch:
          type: string
          example: "main"
        commit_hash:
          type: string
          example: "abc123def456"
    Blueprint:
      type: object
      required:
        - name
        - source
        - runtime
      properties:
        name:
          type: string
          example: "my-app"
        description:
          type: string
          example: "My application description"
        source:
          type: string
          enum: [remote, image]
        runtime:
          $ref: '#/components/schemas/RuntimeObj'
        remote:
          $ref: '#/components/schemas/RemoteObj'
        run_cmd:
          type: string
          example: "npm start"
        build_cmd:
          type: string
          example: "npm run build"
        port:
          type: integer
          minimum: 1025
          maximum: 9999
          example: 3000
        working_dir:
          type: string
          example: "/app"
        static_dir:
          type: string
          example: "/app/dist"
        image:
          type: string
          example: "node:18-alpine"
        env_vars:
          type: object
          additionalProperties:
            type: string
          example:
            NODE_ENV: "production"
            PORT: "3000"
        status:
          type: string
        project_id:
          type: string
    DeploymentStatus:
      type: string
      enum:
        - pending
        - in_progress
        - failed
        - completed
    Deployment:
      type: object
      properties:
        id:
          type: string
          example: "01HN2K3M4N5P6Q7R8S9T0U1V2W"
        user_id:
          type: string
          description: ID of the user who created the deployment
        user_email:
          type: string
          format: email
          description: Email of the user who created the deployment
        config:
          $ref: '#/components/schemas/Blueprint'
        status:
          $ref: '#/components/schemas/DeploymentStatus'
        metadata:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DeployRequest:
      type: object
      required:
        - name
        - user_id
        - source
        - runtime
      properties:
        name:
          type: string
          example: "my-app"
        description:
          type: string
          example: "My application description"
        user_id:
          type: string
        source:
          type: string
          enum: [remote, image]
        runtime:
          $ref: '#/components/schemas/RuntimeObj'
        version:
          type: string
        run_cmd:
          type: string
          example: "npm start"
        build_cmd:
          type: string
          example: "npm run build"
        port:
          type: integer
          minimum: 1
          maximum: 65535
          example: 3000
        working_dir:
          type: string
          example: "/app"
        static_dir:
          type: string
          example: "/app/dist"
        image:
          type: string
          example: "node:18-alpine"
        env_vars:
          type: object
          additionalProperties:
            type: string
        secrets:
          type: object
          additionalProperties:
            type: string
        remote:
          $ref: '#/components/schemas/RemoteObj'
        domain:
          type: string
          example: "myapp.example.com"
        dns_provider:
          type: string
          example: "cloudflare"
    DeployResponse:
      type: object
      properties:
        success:
          type: boolean
        id:
          type: string
          example: "01HN2K3M4N5P6Q7R8S9T0U1V2W"
        name:
          type: string
          example: "my-app"
        created_at:
          type: string
          format: date-time
    Service:
      type: object
      properties:
        id:
          type: string
          example: "01HN2K3M4N5P6Q7R8S9T0U1V2W"
        name:
          type: string
          example: "my-service"
        description:
          type: string
          example: "My service description"
        source:
          type: string
        runtime:
          $ref: '#/components/schemas/Runtime'
        runtime_version:
          type: string
          example: "18.0.0"
        run_cmd:
          type: string
          example: "npm start"
        build_cmd:
          type: string
          example: "npm run build"
        port:
          type: integer
          example: 3000
        working_dir:
          type: string
          example: "/app"
        static_dir:
          type: string
          example: "/app/dist"
        image:
          type: string
          example: "node:18-alpine"
        env_vars:
          type: object
          additionalProperties:
            type: string
        status:
          type: string
        remote:
          type: string
          format: uri
        branch:
          type: string
          example: "main"
        commit_hash:
          type: string
          example: "abc123def456"
        blueprint:
          $ref: '#/components/schemas/Blueprint'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ServiceUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        runtime_version:
          type: string
        run_cmd:
          type: string
        build_cmd:
          type: string
        port:
          type: integer
          minimum: 1
          maximum: 65535
        working_dir:
          type: string
        static_dir:
          type: string
        image:
          type: string
        env_vars:
          type: object
          additionalProperties:
            type: string
        branch:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: "Validation failed"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "name"
              message:
                type: string
                example: "Field is required"
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Unauthorized"
            code: "auth.unauthorized"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Insufficient permissions"
            code: "auth.forbidden"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Resource not found"
            code: "resource.not_found"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
            code: "runtime.internal_server_error"
...